unexport allocationManager IsPodResizeInProgress
